+++ backend/src/main/java/com/woowacourse/kkogkkog/auth/presentation/LoginInterceptor.java
@@ -3,7 +3,7 @@ package com.woowacourse.kkogkkog.auth.presentation;
 import static org.hibernate.validator.internal.metadata.core.ConstraintHelper.PAYLOAD;
 
 import com.woowacourse.kkogkkog.auth.application.JwtTokenProvider;
-import com.woowacourse.kkogkkog.exception.auth.UnauthenticatedTokenException;
+import com.woowacourse.kkogkkog.auth.exception.UnauthenticatedTokenException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.springframework.http.HttpMethod;
