+++ backend/src/main/java/com/woowacourse/kkogkkog/coupon/exception/CouponNotFoundException.java
@@ -1,6 +1,6 @@
 package com.woowacourse.kkogkkog.coupon.exception;
 
-import com.woowacourse.kkogkkog.exception.NotFoundException;
+import com.woowacourse.kkogkkog.common.exception.NotFoundException;
 
 public class CouponNotFoundException extends NotFoundException {
