+++ backend/src/main/java/com/woowacourse/kkogkkog/coupon/exception/SameSenderReceiverException.java
@@ -1,6 +1,6 @@
 package com.woowacourse.kkogkkog.coupon.exception;
 
-import com.woowacourse.kkogkkog.exception.InvalidRequestException;
+import com.woowacourse.kkogkkog.common.exception.InvalidRequestException;
 
 public class SameSenderReceiverException extends InvalidRequestException {
