+++ backend/src/main/java/com/woowacourse/kkogkkog/infrastructure/exception/BotInstallationFailedException.java
@@ -1,4 +1,4 @@
-package com.woowacourse.kkogkkog.exception.infrastructure;
+package com.woowacourse.kkogkkog.infrastructure.exception;
 
 public class BotInstallationFailedException extends RuntimeException {
