+++ backend/src/main/java/com/woowacourse/kkogkkog/member/application/MemberService.java
@@ -15,8 +15,8 @@ import com.woowacourse.kkogkkog.member.domain.WorkspaceUser;
 import com.woowacourse.kkogkkog.member.domain.repository.MemberHistoryRepository;
 import com.woowacourse.kkogkkog.member.domain.repository.MemberRepository;
 import com.woowacourse.kkogkkog.member.domain.repository.WorkspaceUserRepository;
-import com.woowacourse.kkogkkog.exception.member.MemberHistoryNotFoundException;
-import com.woowacourse.kkogkkog.exception.member.MemberNotFoundException;
+import com.woowacourse.kkogkkog.member.exception.MemberHistoryNotFoundException;
+import com.woowacourse.kkogkkog.member.exception.MemberNotFoundException;
 import com.woowacourse.kkogkkog.infrastructure.dto.SlackUserInfo;
 import java.util.List;
 import java.util.Optional;
