+++ backend/src/test/java/com/woowacourse/kkogkkog/reservation/domain/ReservationStatusTest.java
@@ -4,8 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.woowacourse.kkogkkog.coupon.domain.CouponEvent;
-import com.woowacourse.kkogkkog.exception.InvalidRequestException;
-import com.woowacourse.kkogkkog.reservation.domain.ReservationStatus;
+import com.woowacourse.kkogkkog.common.exception.InvalidRequestException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
